+++ __tests__/changelog-updater.test.ts
@@ -18,7 +18,7 @@ const CHANGELOG_WITH_PROPER_SECTIONS_AND_ENTRIES = `# Changelog
 
 ## [v1.0.0]
 ### Dependencies
-- Bumps \`different-package\` from v1 to v2`
+- Bump \`different-package\` from v1 to v2`
 
 test('adds an entry to the changelog when section already exists with entry', async () => {
   mockReadStream(CHANGELOG_WITH_PROPER_SECTIONS_AND_ENTRIES)
@@ -31,15 +31,15 @@ test('adds an entry to the changelog when section already exists with entry', as
 
 ## [v1.0.0]
 ### Dependencies
-- Bumps \`different-package\` from v1 to v2
-- Bumps \`package\` from v1 to v2`)
+- Bump \`different-package\` from v1 to v2
+- Bump \`package\` from v1 to v2`)
 })
 
 const CHANGELOG_WITH_PROPER_SECTIONS_AND_ENTRIES_UNRELEASED = `# Changelog
 
 ## [UNRELEASED]
 ### Dependencies
-- Bumps \`different-package\` from v1 to v2`
+- Bump \`different-package\` from v1 to v2`
 
 test('adds an entry to the changelog when section exists under default unreleased version', async () => {
   mockReadStream(CHANGELOG_WITH_PROPER_SECTIONS_AND_ENTRIES_UNRELEASED)
@@ -52,8 +52,8 @@ test('adds an entry to the changelog when section exists under default unrelease
 
 ## [UNRELEASED]
 ### Dependencies
-- Bumps \`different-package\` from v1 to v2
-- Bumps \`package\` from v1 to v2`)
+- Bump \`different-package\` from v1 to v2
+- Bump \`package\` from v1 to v2`)
 })
 
 const CHANGELOG_WITH_PROPER_SECTIONS_UNRELEASED = `# Changelog
@@ -72,7 +72,7 @@ test('adds an entry to the changelog when section already exists, but no entry',
 
 ## [UNRELEASED]
 ### Dependencies
-- Bumps \`package\` from v1 to v2`)
+- Bump \`package\` from v1 to v2`)
 })
 
 const CHANGELOG_MISSING_DEPENDECIES = `# Changelog
@@ -92,18 +92,18 @@ test('adds an entry to the changelog when version exists but section does not', 
 
 ## [UNRELEASED]
 ### Dependencies
-- Bumps \`package\` from v1 to v2`)
+- Bump \`package\` from v1 to v2`)
 })
 
 const CHANGELOG_WITH_MULTIPLE_VERSIONS = `# Changelog
 
 ## [UNRELEASED]
 ### Dependencies
-- Bumps \`different-package\` from v1 to v2
+- Bump \`different-package\` from v1 to v2
 
 ## [v1.0.0]
 ### Dependencies
-- Bumps \`foo\` from bar to foo-bar
+- Bump \`foo\` from bar to foo-bar
 `
 
 test('adds an entry to the changelog - multiple versions', async () => {
@@ -117,12 +117,12 @@ test('adds an entry to the changelog - multiple versions', async () => {
 
 ## [UNRELEASED]
 ### Dependencies
-- Bumps \`different-package\` from v1 to v2
-- Bumps \`package\` from v1 to v2
+- Bump \`different-package\` from v1 to v2
+- Bump \`package\` from v1 to v2
 
 ## [v1.0.0]
 ### Dependencies
-- Bumps \`foo\` from bar to foo-bar`)
+- Bump \`foo\` from bar to foo-bar`)
 })
 
 const CHANGELOG_WITH_NO_VERSION = `# Changelog
@@ -143,7 +143,7 @@ const CHANGELOG_WITH_DUPLICATE_ENTRY = `# Changelog
 
 ## [v1.0.0]
 ### Dependencies
-- Bumps \`package\` from v1 to v2`
+- Bump \`package\` from v1 to v2`
 
 test('does not update the changelog on duplicate entry', async () => {
   mockReadStream(CHANGELOG_WITH_DUPLICATE_ENTRY)
@@ -157,7 +157,7 @@ const CHANGELOG_WITH_DUPLICATE_ENTRY_NOT_LAST_LINE = `# Changelog
 
 ## [v1.0.0]
 ### Dependencies
-- Bumps \`package\` from v1 to v2
+- Bump \`package\` from v1 to v2
 
 ## [v0.9.0]
 ### Added
@@ -175,7 +175,7 @@ const CHANGELOG_WITH_ENTRY_TO_UPDATE = `# Changelog
 
 ## [v1.0.0]
 ### Dependencies
-- Bumps \`package\` from v1 to v1.1`
+- Bump \`package\` from v1 to v1.1`
 
 test('updates an entry for an existing package in the same version', async () => {
   mockReadStream(CHANGELOG_WITH_ENTRY_TO_UPDATE)
@@ -189,7 +189,7 @@ test('updates an entry for an existing package in the same version', async () =>
 
 ## [v1.0.0]
 ### Dependencies
-- Bumps \`package\` from v1 to v2`
+- Bump \`package\` from v1 to v2`
   )
 })
 
@@ -213,7 +213,7 @@ test('updates version with new section and entry', async () => {
 ### Added
 ### Removed
 ### Dependencies
-- Bumps \`package\` from v1 to v2`
+- Bump \`package\` from v1 to v2`
   )
 })
 
@@ -221,11 +221,11 @@ const CHANGELOG_WITH_MULTI_VERSION_PACKAGE_UPDATES = `# Changelog
 
 ## [v1.0.0]
 ### Dependencies
-- Bumps \`package\` from v1 to v1.1
+- Bump \`package\` from v1 to v1.1
 
 ## [v0.9.0]
 ### Dependencies
-- Bumps \`package\` from alpha to v1`
+- Bump \`package\` from alpha to v1`
 
 test('Does not update lines additional times', async () => {
   mockReadStream(CHANGELOG_WITH_MULTI_VERSION_PACKAGE_UPDATES)
@@ -239,11 +239,11 @@ test('Does not update lines additional times', async () => {
 
 ## [v1.0.0]
 ### Dependencies
-- Bumps \`package\` from v1 to v2
+- Bump \`package\` from v1 to v2
 
 ## [v0.9.0]
 ### Dependencies
-- Bumps \`package\` from alpha to v1`
+- Bump \`package\` from alpha to v1`
   )
 })
 
@@ -255,12 +255,12 @@ const CHANGELOG_WITH_EXISTING_SECTION_AND_SEPARATED_SECTIONS = `# Changelog
 - Added a new feature
 
 ### Dependencies
-- Bumps \`other-package\` from v2 to v3
+- Bump \`other-package\` from v2 to v3
 
 ## [v0.9.0]
 
 ### Dependencies
-- Bumps \`package\` from alpha to v1`
+- Bump \`package\` from alpha to v1`
 
 test('Updates existing section when sections separated by blank lines', async () => {
   mockReadStream(CHANGELOG_WITH_EXISTING_SECTION_AND_SEPARATED_SECTIONS)
@@ -278,13 +278,13 @@ test('Updates existing section when sections separated by blank lines', async ()
 - Added a new feature
 
 ### Dependencies
-- Bumps \`other-package\` from v2 to v3
-- Bumps \`package\` from v1 to v2
+- Bump \`other-package\` from v2 to v3
+- Bump \`package\` from v1 to v2
 
 ## [v0.9.0]
 
 ### Dependencies
-- Bumps \`package\` from alpha to v1`
+- Bump \`package\` from alpha to v1`
   )
 })
 
@@ -301,7 +301,7 @@ const CHANGELOG_WITHOUT_EXISTING_SECTION_AND_SEPARATED_SECTIONS = `# Changelog
 ## [v0.9.0]
 
 ### Dependencies
-- Bumps \`package\` from alpha to v1`
+- Bump \`package\` from alpha to v1`
 
 test('Adds section when sections separated by blank lines', async () => {
   mockReadStream(CHANGELOG_WITHOUT_EXISTING_SECTION_AND_SEPARATED_SECTIONS)
@@ -321,12 +321,12 @@ test('Adds section when sections separated by blank lines', async () => {
 ### Removed
 - Removed a feature
 ### Dependencies
-- Bumps \`package\` from v1 to v2
+- Bump \`package\` from v1 to v2
 
 ## [v0.9.0]
 
 ### Dependencies
-- Bumps \`package\` from alpha to v1`
+- Bump \`package\` from alpha to v1`
   )
 })
 
@@ -338,7 +338,7 @@ const CHANGELOG_WITH_EXISTING_SECTION_BETWEEN_OTHERS = `# Changelog
 - Added a new feature
 
 ### Dependencies
-- Bumps \`other-package\` from v2 to v3
+- Bump \`other-package\` from v2 to v3
 
 ### Removed
 - Removed a feature
@@ -346,7 +346,7 @@ const CHANGELOG_WITH_EXISTING_SECTION_BETWEEN_OTHERS = `# Changelog
 ## [v0.9.0]
 
 ### Dependencies
-- Bumps \`package\` from alpha to v1`
+- Bump \`package\` from alpha to v1`
 
 test('Updates existing section when between other sections', async () => {
   mockReadStream(CHANGELOG_WITH_EXISTING_SECTION_BETWEEN_OTHERS)
@@ -364,8 +364,8 @@ test('Updates existing section when between other sections', async () => {
 - Added a new feature
 
 ### Dependencies
-- Bumps \`other-package\` from v2 to v3
-- Bumps \`package\` from v1 to v2
+- Bump \`other-package\` from v2 to v3
+- Bump \`package\` from v1 to v2
 
 ### Removed
 - Removed a feature
@@ -373,7 +373,7 @@ test('Updates existing section when between other sections', async () => {
 ## [v0.9.0]
 
 ### Dependencies
-- Bumps \`package\` from alpha to v1`
+- Bump \`package\` from alpha to v1`
   )
 })